requirements.txt
Django>=5.2.3,<5.3
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt>=5.3.0,<5.4
psycopg2-binary>=2.9.9,<2.10
Pillow>=10.1.0,<10.2
python-dotenv>=1.0.0,<1.1